@@ -1,33 +1,81 @@
+"use client";
+
 import React, { useCallback, useRef, useState } from "react";
 import Image from "next/image";
 import Link from "next/link";
-import { useRouter } from "next/router";
+import { usePathname } from "next/navigation";
 import {
   Bars3CenterLeftIcon,
   Bars3Icon,
   BugAntIcon,
   CheckBadgeIcon,
-  MagnifyingGlassIcon,
   PencilIcon,
   PlusCircleIcon,
 } from "@heroicons/react/24/outline";
 import { FaucetButton, RainbowKitCustomConnectButton } from "~~/components/scaffold-eth";
 import { useOutsideClick } from "~~/hooks/scaffold-eth";
 
-const NavLink = ({ href, children }: { href: string; children: React.ReactNode }) => {
-  const router = useRouter();
-  const isActive = router.pathname === href;
+type HeaderMenuLink = {
+  label: string;
+  href: string;
+  icon?: React.ReactNode;
+};
+
+export const menuLinks: HeaderMenuLink[] = [
+  {
+    label: "Home",
+    href: "/",
+  },
+  {
+    label: "Multisig",
+    href: "/multisig",
+    icon: <PencilIcon className="h-4 w-4" />,
+  },
+  {
+    label: "Owners",
+    href: "/owners",
+    icon: <CheckBadgeIcon className="h-4 w-4" />,
+  },
+  {
+    label: "Create",
+    href: "/create",
+    icon: <PlusCircleIcon className="h-4 w-4" />,
+  },
+  {
+    label: "Pool",
+    href: "/pool",
+    icon: <Bars3CenterLeftIcon className="h-4 w-4" />,
+  },
+  {
+    label: "Debug Contracts",
+    href: "/debug",
+    icon: <BugAntIcon className="h-4 w-4" />,
+  },
+];
+
+export const HeaderMenuLinks = () => {
+  const pathname = usePathname();
 
   return (
-    <Link
-      href={href}
-      passHref
-      className={`${
-        isActive ? "bg-secondary shadow-md" : ""
-      } hover:bg-secondary hover:shadow-md focus:!bg-secondary py-1.5 px-3 text-sm rounded-full gap-2`}
-    >
-      {children}
-    </Link>
+    <>
+      {menuLinks.map(({ label, href, icon }) => {
+        const isActive = pathname === href;
+        return (
+          <li key={href}>
+            <Link
+              href={href}
+              passHref
+              className={`${
+                isActive ? "bg-secondary shadow-md" : ""
+              } hover:bg-secondary hover:shadow-md focus:!bg-secondary active:!text-neutral py-1.5 px-3 text-sm rounded-full gap-2 grid grid-flow-col`}
+            >
+              {icon}
+              <span>{label}</span>
+            </Link>
+          </li>
+        );
+      })}
+    </>
   );
 };
 
@@ -40,47 +88,6 @@ export const Header = () => {
   useOutsideClick(
     burgerMenuRef,
     useCallback(() => setIsDrawerOpen(false), []),
-  );
-
-  const navLinks = (
-    <>
-      <li>
-        <NavLink href="/multisig">
-          <PencilIcon className="h-4 w-4" />
-          Multisig
-        </NavLink>
-      </li>
-      <li>
-        <NavLink href="/owners">
-          <CheckBadgeIcon className="h-4 w-4" />
-          Owners
-        </NavLink>
-      </li>
-      <li>
-        <NavLink href="/create">
-          <PlusCircleIcon className="h-4 w-4" />
-          Create
-        </NavLink>
-      </li>
-      <li>
-        <NavLink href="/pool">
-          <Bars3CenterLeftIcon className="h-4 w-4" />
-          Pool
-        </NavLink>
-      </li>
-      <li>
-        <NavLink href="/debug">
-          <BugAntIcon className="h-4 w-4" />
-          Debug Contracts
-        </NavLink>
-      </li>
-      <li>
-        <NavLink href="/blockexplorer">
-          <MagnifyingGlassIcon className="h-4 w-4" />
-          Block Explorer
-        </NavLink>
-      </li>
-    </>
   );
 
   return (
@@ -99,12 +106,12 @@ export const Header = () => {
           {isDrawerOpen && (
             <ul
               tabIndex={0}
-              className="menu menu-compact dropdown-content mt-3 p-2 shadow bg-primary rounded-box w-52"
+              className="menu menu-compact dropdown-content mt-3 p-2 shadow bg-base-100 rounded-box w-52"
               onClick={() => {
                 setIsDrawerOpen(false);
               }}
             >
-              {navLinks}
+              <HeaderMenuLinks />
             </ul>
           )}
         </div>
@@ -117,7 +124,9 @@ export const Header = () => {
             <span className="text-xs">#6: Multisig Wallet</span>
           </div>
         </Link>
-        <ul className="hidden lg:flex lg:flex-nowrap menu menu-horizontal px-1 gap-2">{navLinks}</ul>
+        <ul className="hidden lg:flex lg:flex-nowrap menu menu-horizontal px-1 gap-2">
+          <HeaderMenuLinks />
+        </ul>
       </div>
       <div className="navbar-end flex-grow mr-4">
         <RainbowKitCustomConnectButton />
