@@ -3,7 +3,7 @@ import { CommonInputProps, InputBase, IntegerVariant, isValidInteger } from "~~/
 
 type IntegerInputProps = CommonInputProps<string | bigint> & {
   variant?: IntegerVariant;
-  hideSuffix?: boolean;
+  disableMultiplyBy1e18?: boolean;
 };
 
 export const IntegerInput = ({
@@ -13,7 +13,7 @@ export const IntegerInput = ({
   placeholder,
   disabled,
   variant = IntegerVariant.UINT256,
-  hideSuffix,
+  disableMultiplyBy1e18 = false,
 }: IntegerInputProps) => {
   const [inputError, setInputError] = useState(false);
   const multiplyBy1e18 = useCallback(() => {
@@ -43,8 +43,8 @@ export const IntegerInput = ({
       onChange={onChange}
       disabled={disabled}
       suffix={
-        !hideSuffix &&
-        !inputError && (
+        !inputError &&
+        !disableMultiplyBy1e18 && (
           <div
             className="space-x-4 flex tooltip tooltip-top tooltip-secondary before:content-[attr(data-tip)] before:right-[-10px] before:left-auto before:transform-none"
             data-tip="Multiply by 10^18 (wei)"
